move:#27 닉네임 가져오기 서버 액션으로 이동
