+++ src/services/signIn.ts
@@ -18,3 +18,14 @@ export const signInSupabase = async ({ email, password }: SignInProps) => {
   }
   return data;
 };
+
+// supabase에서 해당 email user의 nickname 가져오기
+export const getUserByEmail = async (email: string) => {
+  //single(): 하나의 행만 가져옴 (select만 할 시 배열)
+  const { data, error } = await supabase.from('users').select('nickname').eq('email', email).single();
+  if (error) {
+    console.error('유저 조회 실패', error.message);
+    return null;
+  }
+  return data;
+};
